refactor: move worktree config flag to utils
