+++ src/tools/delegation/inputFields.ts
@@ -9,13 +9,13 @@ import { z } from 'zod';
 import type { ToolResult } from '@shared/schemas/toolResult';
 import { formatBytes } from '@shared/utils/string';
 import { parseWorkingDirectory } from '@tools/pathResolution';
-import { isWorktreeSupportEnabled } from '@tools/worktreeConfig';
 
 // Local imports - memory
 import { displayToStoragePath } from '@tools/memory/memoryUtils';
 
 // Local imports - utils
 import { AbsoluteFS, WorkspaceFS } from '@utils/files';
+import { isWorktreeSupportEnabled } from '@utils/config/worktreeConfig';
 import {
   extractErrorMessage,
   toErrorMessage,
